removi o "final.c"; alterei o "expressão.c" e adicionei o "#include "expressao.c"" no main.c.
